requirements.txt
typer>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
hypothesis>=6.92.0
pytest>=7.4.0
